to_usd placed in __init__
import into shopping and robo
