+++ app/__init__.py
@@ -6,3 +6,13 @@
 # ... facilitates local imports from one file to another,
 # ... including importing app code for testing purposes
 #
+
+def to_usd(my_price):
+    """
+    Formats a number as currency (USD) with dollar sign, two decimals, and thousands separator
+    
+    Params: my_price (numeric, int or float) to be formatted
+    
+    Examples: to_usd(21.645)
+    """
+    return f"${my_price:,.2f}"
