+++ app/shopping.py
@@ -2,16 +2,7 @@ import os
 from datetime import datetime
 from pandas import read_csv
 
-
-def to_usd(my_price):
-    """
-    Formats a number as currency (USD) with dollar sign, two decimals, and thousands separator
-    
-    Params: my_price (numeric, int or float) to be formatted
-    
-    Examples: to_usd(21.645)
-    """
-    return f"${my_price:,.2f}"
+from app.__init__ import to_usd
 
 
 def lookup_product(product_id, all_products):
